Decrease TTFB by adding separate page for sharables
